moved code from translations dir to service
