+++ Resources/config/services.yml
@@ -1,12 +1,12 @@
 services:
   ongr_translations.translation_manager:
-    class: ONGR\TranslationsBundle\Translation\TranslationManager
+    class: ONGR\TranslationsBundle\Service\TranslationManager
     arguments:
       - "@ongr_translations.translation_repository"
       - "@ongr_translations.history_manager"
       - "@event_dispatcher"
   ongr_translations.history_manager:
-    class: ONGR\TranslationsBundle\Translation\HistoryManager
+    class: ONGR\TranslationsBundle\Service\HistoryManager
     arguments:
       - "@ongr_translations.history_repository"
   ongr_translations.command.export:
@@ -15,12 +15,12 @@ services:
       - [setContainer, ["@service_container"]]
 
   ongr_translations.file_import:
-    class: ONGR\TranslationsBundle\Translation\Import\FileImport
+    class: ONGR\TranslationsBundle\Service\Import\FileImport
     arguments:
       - [ "@ongr_translations.loaders_container" ]
 
   ongr_translations.import:
-    class: ONGR\TranslationsBundle\Service\Import
+    class: ONGR\TranslationsBundle\Service\Import\Import
     arguments:
       - "@ongr_translations.file_import"
       - "@es.manager"
@@ -33,10 +33,10 @@ services:
       - [ setFormats, [ %ongr_translations.formats% ] ]
 
   ongr_translations.export.yaml:
-    class: ONGR\TranslationsBundle\Translation\Export\YmlExport
+    class: ONGR\TranslationsBundle\Service\Export\YmlExport
 
   ongr_translations.export:
-    class: ONGR\TranslationsBundle\Service\Export
+    class: ONGR\TranslationsBundle\Service\Export\Export
     arguments:
       - "@ongr_translations.loaders_container"
       - "@ongr_translations.translation_repository"
