+++ Service/Export/ExporterInterface.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace ONGR\TranslationsBundle\Translation\Export;
+namespace ONGR\TranslationsBundle\Service\Export;
 
 /**
  * This interface should be implemented by file exporters.
